Create initial validiac fork
This commit introduces an initial fork of validkube that provides
similar capabilities for Terraform HCLs and related files.

Four capabilities are provided:
- lint: lints HCLs using tflint
- secure: checks for security issues using tfsec
- cost: calculates costs using infracost
- map: generates graph using inframap

The four dependencies are automatically downloaded using `make deps`.
Their versions are defined in the Makefile.

When the validiac binary is built, the location of the dependencies is
injected using Go build-time variables.

The binary can function both as an AWS Lambda function and as a simple
command line interface.

The UI has not been modified yet, and more work is required on the
backend:

- Infracost requires an API key for usage.
- When building a lambda, injected dependency paths should probably be
  different than when build for local execution.
- The tools mostly make sense when they can work on complete directories
  of Terraform files rather than individual files. This is not addressed
  yet.